add forecast, key_levels
