+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = technical_analysis
-version = 0.0.3
+version = 0.0.4
 url = https://github.com/trevormcguire/technical-analysis
 author = Trevor McGuire
 author_email = trevormcguire13@gmail.com
@@ -22,6 +22,7 @@ install_requires =
     numpy
     mplfinance
     pandas
+    scipy
 
 [options.packages.find]
 where=src
