+++ src/technical_analysis/key_levels/support_resistance.py
@@ -1,0 +1,13 @@
+import numpy as np
+import pandas as pd
+from scipy.signal import argrelextrema
+
+
+
+class Auto_SR(object):
+    """
+    Automatic Identification of Support and Resistance Levels
+    """
+    def __init__(self):
+        pass
+
