+++ src/technical_analysis/stats/__init__.py
@@ -1,1 +1,2 @@
 from technical_analysis.stats.stats import *
+from technical_analysis.stats.hurst import *
